requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cleanup==7.0.0
django-extensions==3.2.1
Pillow==9.5.0
psycopg2==2.9.6
sqlparse==0.4.4
